More fixes and separation between core and package tests
