+++ .circleci/config.yml
@@ -96,15 +96,14 @@ jobs:
 
       - save_cache:
           paths:
-            - ./emsdk/emsdk
             - ~/.ccache
-          key: v1-emsdk-{{ checksum "emsdk/Makefile" }}-v20201203-{{ .BuildNum }}
+          key: v1-emsdk-{{ checksum "emsdk/Makefile" }}-v20201205-{{ .BuildNum }}
 
       - persist_to_workspace:
           root: .
           paths:
+            - ./packages/.artifacts
             - ./build
-            - ./cpython/build/3.8.2/host
 
       - store_artifacts:
           path: /home/circleci/repo/build/
@@ -117,7 +116,8 @@ jobs:
       - run:
           name: test
           command: |
-            pip install pytest-httpserver
+            # distlib is required by micropip
+            pip install distlib
             pytest src packages/micropip/ -v -k firefox
 
   test-core-chrome:
@@ -128,7 +128,8 @@ jobs:
       - run:
           name: test
           command: |
-            pip install pytest-httpserver
+            # distlib is required by micropip
+            pip install distlib
             pytest src packages/micropip/ -v -k chrome
 
   test-packages-firefox:
