+++ src/tests/test_webworker.py
@@ -1,22 +1,11 @@
 import pytest
 
 
-def test_runwebworker(selenium_standalone):
-    output = selenium_standalone.run_webworker(
-        """
-        import numpy as np
-        x = np.zeros(5)
-        str(x)
-        """
-    )
-    assert output == "[0. 0. 0. 0. 0.]"
-
-
 def test_runwebworker_different_package_name(selenium_standalone):
     output = selenium_standalone.run_webworker(
         """
-        import dateutil
-        dateutil.__version__
+        import pyparsing
+        pyparsing.__version__
         """
     )
     assert isinstance(output, str)
@@ -56,7 +45,7 @@ def test_runwebworker_exception_after_import(selenium_standalone):
     with pytest.raises(selenium_standalone.JavascriptException, match=msg):
         selenium_standalone.run_webworker(
             """
-            import numpy as np
+            import pyparsing
             x = np.empty(5)
             42 / 0
             """
